feat(rust): update and refactor many D-Bus proxies
