+++ rust/agama-lib/src/network/proxies.rs
@@ -21,9 +21,9 @@
 //! D-Bus interface proxies for: `org.opensuse.Agama*.**.*`
 //!
 //! This code was generated by `zbus-xmlgen` `3.1.0` from DBus introspection data.`.
-use zbus::dbus_proxy;
+use zbus::proxy;
 
-#[dbus_proxy(
+#[proxy(
     interface = "org.opensuse.Agama1.Network.Devices",
     default_service = "org.opensuse.Agama1",
     default_path = "/org/opensuse/Agama1/Network/devices"
@@ -33,24 +33,24 @@ trait Devices {
     fn get_devices(&self) -> zbus::Result<Vec<zbus::zvariant::OwnedObjectPath>>;
 }
 
-#[dbus_proxy(
+#[proxy(
     interface = "org.opensuse.Agama1.Network.Device",
     default_service = "org.opensuse.Agama1",
     default_path = "/org/opensuse/Agama1/Network/devices/1"
 )]
 trait Device {
     /// Name property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn name(&self) -> zbus::Result<String>;
     /// Type property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn type_(&self) -> zbus::Result<u8>;
     /// State property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn state(&self) -> zbus::Result<u8>;
 }
 
-#[dbus_proxy(
+#[proxy(
     interface = "org.opensuse.Agama1.Network.Connections",
     default_service = "org.opensuse.Agama1",
     default_path = "/org/opensuse/Agama1/Network/connections"
@@ -75,11 +75,11 @@ trait Connections {
     fn remove_connection(&self, uuid: &str) -> zbus::Result<()>;
 
     /// ConnectionAdded signal
-    #[dbus_proxy(signal)]
+    #[zbus(signal)]
     fn connection_added(&self, id: &str, path: zbus::zvariant::ObjectPath<'_>) -> zbus::Result<()>;
 }
 
-#[dbus_proxy(
+#[proxy(
     interface = "org.opensuse.Agama1.Network.Connection.Wireless",
     default_service = "org.opensuse.Agama1",
     default_path = "/org/opensuse/Agama1/Network"
@@ -88,160 +88,160 @@ trait Wireless {
     /// Returns the operating mode of the Wireless device
     ///
     /// Possible values are 'unknown', 'adhoc', 'infrastructure', 'ap' or 'mesh'
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn mode(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_mode(&self, value: &str) -> zbus::Result<()>;
 
     /// Password property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn password(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_password(&self, value: &str) -> zbus::Result<()>;
 
     /// SSID property
-    #[dbus_proxy(property, name = "SSID")]
+    #[zbus(property, name = "SSID")]
     fn ssid(&self) -> zbus::Result<Vec<u8>>;
-    #[dbus_proxy(property, name = "SSID")]
+    #[zbus(property, name = "SSID")]
     fn set_ssid(&self, value: &[u8]) -> zbus::Result<()>;
 
     /// Wireless Security property
     ///
     /// Possible values are 'none', 'owe', 'ieee8021x', 'wpa-psk', 'sae',
     ///     'wpa-eap', 'wpa-eap-suite-b-192'
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn security(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_security(&self, value: &str) -> zbus::Result<()>;
 }
 
-#[dbus_proxy(
+#[proxy(
     interface = "org.opensuse.Agama1.Network.Connection",
     default_service = "org.opensuse.Agama1",
     default_path = "/org/opensuse/Agama1/Network"
 )]
 trait Connection {
     /// Id property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn id(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn interface(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_interface(&self, interface: &str) -> zbus::Result<()>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn mac_address(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_mac_address(&self, mac_address: &str) -> zbus::Result<()>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn mtu(&self) -> zbus::Result<u32>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_mtu(&self, mtu: u32) -> zbus::Result<()>;
 }
 
-#[dbus_proxy(
+#[proxy(
     interface = "org.opensuse.Agama1.Network.Connection.IP",
     default_service = "org.opensuse.Agama1",
     default_path = "/org/opensuse/Agama1/Network/connections/0"
 )]
 trait IP {
     /// Addresses property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn addresses(&self) -> zbus::Result<Vec<String>>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_addresses(&self, value: &[&str]) -> zbus::Result<()>;
 
     /// Gateway4 property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn gateway4(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_gateway4(&self, value: &str) -> zbus::Result<()>;
 
     /// Gateway6 property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn gateway6(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_gateway6(&self, value: &str) -> zbus::Result<()>;
 
     /// Method4 property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn method4(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_method4(&self, value: &str) -> zbus::Result<()>;
 
     /// Method6 property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn method6(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_method6(&self, value: &str) -> zbus::Result<()>;
 
     /// Nameservers property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn nameservers(&self) -> zbus::Result<Vec<String>>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_nameservers(&self, value: &[&str]) -> zbus::Result<()>;
 
     /// DNS searchlist property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn dns_searchlist(&self) -> zbus::Result<Vec<String>>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_dns_searchlist(&self, value: &[&str]) -> zbus::Result<()>;
 
     /// Ignore auto DNS property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn ignore_auto_dns(&self) -> zbus::Result<bool>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_ignore_auto_dns(&self, value: bool) -> zbus::Result<()>;
 }
 
-#[dbus_proxy(
+#[proxy(
     interface = "org.opensuse.Agama1.Network.Connection.Match",
     default_service = "org.opensuse.Agama1",
     default_path = "/org/opensuse/Agama1/Network"
 )]
 trait Match {
     /// Driver property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn driver(&self) -> zbus::Result<Vec<String>>;
     fn set_driver(&self, value: &[&str]) -> zbus::Result<()>;
 
     /// Interface property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn interface(&self) -> zbus::Result<Vec<String>>;
     fn set_interface(&self, value: &[&str]) -> zbus::Result<()>;
 
     /// Path property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn path(&self) -> zbus::Result<Vec<String>>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_path(&self, value: &[&str]) -> zbus::Result<()>;
 
     /// Path property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn kernel(&self) -> zbus::Result<Vec<String>>;
     fn set_kernel(&self, value: &[&str]) -> zbus::Result<()>;
 }
 
-#[dbus_proxy(
+#[proxy(
     interface = "org.opensuse.Agama1.Network.Connection.Bond",
     default_service = "org.opensuse.Agama1",
     default_path = "/org/opensuse/Agama1/Network"
 )]
 trait Bond {
     /// Mode property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn mode(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_mode(&self, value: &str) -> zbus::Result<()>;
 
     /// Ports property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn options(&self) -> zbus::Result<String>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_options(&self, value: &str) -> zbus::Result<()>;
 
     /// Ports property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn ports(&self) -> zbus::Result<Vec<String>>;
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn set_ports(&self, value: &[&str]) -> zbus::Result<()>;
 }
