+++ rust/agama-lib/src/product/proxies.rs
@@ -18,17 +18,34 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! # DBus interface proxy for: `org.opensuse.Agama1.Registration`
+//! # D-Bus interface proxy for: `org.opensuse.Agama1.Registration`
 //!
-//! This code was generated by `zbus-xmlgen` `3.1.1` from DBus introspection data.
-use zbus::dbus_proxy;
-
-#[dbus_proxy(
-    interface = "org.opensuse.Agama1.Registration",
+//! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
+//! Source: `org.opensuse.Agama.Software1.Product.bus.xml`.
+//!
+//! You may prefer to adapt it, instead of using it verbatim.
+//!
+//! More information can be found in the [Writing a client proxy] section of the zbus
+//! documentation.
+//!
+//! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
+//! following zbus API can be used:
+//!
+//! * [`zbus::fdo::PropertiesProxy`]
+//! * [`zbus::fdo::IntrospectableProxy`]
+//!
+//! Consequently `zbus-xmlgen` did not generate code for the above interfaces.
+//!
+//! [Writing a client proxy]: https://dbus2.github.io/zbus/client.html
+//! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
+use zbus::proxy;
+#[proxy(
     default_service = "org.opensuse.Agama.Software1",
-    default_path = "/org/opensuse/Agama/Software1/Product"
+    default_path = "/org/opensuse/Agama/Software1/Product",
+    interface = "org.opensuse.Agama1.Registration",
+    assume_defaults = true
 )]
-trait Registration {
+pub trait Registration {
     /// Deregister method
     fn deregister(&self) -> zbus::Result<(u32, String)>;
 
@@ -36,18 +53,18 @@ trait Registration {
     fn register(
         &self,
         reg_code: &str,
-        options: std::collections::HashMap<&str, zbus::zvariant::Value<'_>>,
+        options: std::collections::HashMap<&str, &zbus::zvariant::Value<'_>>,
     ) -> zbus::Result<(u32, String)>;
 
     /// Email property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn email(&self) -> zbus::Result<String>;
 
     /// RegCode property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn reg_code(&self) -> zbus::Result<String>;
 
     /// Requirement property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn requirement(&self) -> zbus::Result<u32>;
 }
