+++ rust/agama-lib/src/proxies/manager1.rs
@@ -1,0 +1,61 @@
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Manager1`
+//!
+//! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
+//! Source: `org.opensuse.Agama1.Manager.bus.xml`.
+//!
+//! You may prefer to adapt it, instead of using it verbatim.
+//!
+//! More information can be found in the [Writing a client proxy] section of the zbus
+//! documentation.
+//!
+//! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
+//! following zbus API can be used:
+//!
+//! * [`zbus::fdo::PropertiesProxy`]
+//! * [`zbus::fdo::IntrospectableProxy`]
+//!
+//! Consequently `zbus-xmlgen` did not generate code for the above interfaces.
+//!
+//! [Writing a client proxy]: https://dbus2.github.io/zbus/client.html
+//! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
+use zbus::proxy;
+#[proxy(
+    default_service = "org.opensuse.Agama.Manager1",
+    default_path = "/org/opensuse/Agama/Manager1",
+    interface = "org.opensuse.Agama.Manager1",
+    assume_defaults = true
+)]
+pub trait Manager1 {
+    /// CanInstall method
+    fn can_install(&self) -> zbus::Result<bool>;
+
+    /// CollectLogs method
+    fn collect_logs(&self) -> zbus::Result<String>;
+
+    /// Commit method
+    fn commit(&self) -> zbus::Result<()>;
+
+    /// Finish method
+    fn finish(&self) -> zbus::Result<()>;
+
+    /// Probe method
+    fn probe(&self) -> zbus::Result<()>;
+
+    /// BusyServices property
+    #[zbus(property)]
+    fn busy_services(&self) -> zbus::Result<Vec<String>>;
+
+    /// CurrentInstallationPhase property
+    #[zbus(property)]
+    fn current_installation_phase(&self) -> zbus::Result<u32>;
+
+    /// IguanaBackend property
+    #[zbus(property)]
+    fn iguana_backend(&self) -> zbus::Result<bool>;
+
+    /// InstallationPhases property
+    #[zbus(property)]
+    fn installation_phases(
+        &self,
+    ) -> zbus::Result<Vec<std::collections::HashMap<String, zbus::zvariant::OwnedValue>>>;
+}
