+++ rust/agama-lib/src/proxies/questions.rs
@@ -18,11 +18,11 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-mod software;
-pub use software::Software1Proxy;
+mod questions;
+pub use questions::QuestionsProxy;
 
-mod product;
-pub use product::{Product, ProductProxy as SoftwareProductProxy};
+mod generic;
+pub use generic::GenericProxy as GenericQuestionProxy;
 
-mod proposal;
-pub use proposal::ProposalProxy;
+mod with_password;
+pub use with_password::WithPasswordProxy as QuestionWithPasswordProxy;
