+++ rust/agama-lib/src/proxies/questions/questions.rs
@@ -18,10 +18,10 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! # D-Bus interface proxy for: `org.opensuse.Agama1.Locale`
+//! # D-Bus interface proxy for: `org.opensuse.Agama1.Questions`
 //!
 //! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
-//! Source: `org.opensuse.Agama1.Locale.bus.xml`.
+//! Source: `org.opensuse.Agama1.Questions.bus.xml`.
 //!
 //! You may prefer to adapt it, instead of using it verbatim.
 //!
@@ -31,9 +31,10 @@
 //! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
 //! following zbus API can be used:
 //!
+//! * [`zbus::fdo::IntrospectableProxy`]
+//! * [`zbus::fdo::ObjectManagerProxy`]
 //! * [`zbus::fdo::PeerProxy`]
 //! * [`zbus::fdo::PropertiesProxy`]
-//! * [`zbus::fdo::IntrospectableProxy`]
 //!
 //! Consequently `zbus-xmlgen` did not generate code for the above interfaces.
 //!
@@ -42,34 +43,40 @@
 use zbus::proxy;
 #[proxy(
     default_service = "org.opensuse.Agama1",
-    interface = "org.opensuse.Agama1.Locale",
+    interface = "org.opensuse.Agama1.Questions",
     assume_defaults = true
 )]
-pub trait Locale {
-    /// Commit method
-    fn commit(&self) -> zbus::Result<()>;
+pub trait Questions {
+    /// AddAnswerFile method
+    fn add_answer_file(&self, path: &str) -> zbus::Result<()>;
 
-    /// Keymap property
-    #[zbus(property)]
-    fn keymap(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_keymap(&self, value: &str) -> zbus::Result<()>;
+    /// Delete method
+    fn delete(&self, question: &zbus::zvariant::ObjectPath<'_>) -> zbus::Result<()>;
 
-    /// Locales property
-    #[zbus(property)]
-    fn locales(&self) -> zbus::Result<Vec<String>>;
-    #[zbus(property)]
-    fn set_locales(&self, value: &[&str]) -> zbus::Result<()>;
+    /// New method
+    #[zbus(name = "New")]
+    fn new_question(
+        &self,
+        class: &str,
+        text: &str,
+        options: &[&str],
+        default_option: &str,
+        data: std::collections::HashMap<&str, &str>,
+    ) -> zbus::Result<zbus::zvariant::OwnedObjectPath>;
 
-    /// Timezone property
-    #[zbus(property)]
-    fn timezone(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_timezone(&self, value: &str) -> zbus::Result<()>;
+    /// NewWithPassword method
+    fn new_with_password(
+        &self,
+        class: &str,
+        text: &str,
+        options: &[&str],
+        default_option: &str,
+        data: std::collections::HashMap<&str, &str>,
+    ) -> zbus::Result<zbus::zvariant::OwnedObjectPath>;
 
-    /// UILocale property
-    #[zbus(property, name = "UILocale")]
-    fn uilocale(&self) -> zbus::Result<String>;
-    #[zbus(property, name = "UILocale")]
-    fn set_uilocale(&self, value: &str) -> zbus::Result<()>;
+    /// Interactive property
+    #[zbus(property)]
+    fn interactive(&self) -> zbus::Result<bool>;
+    #[zbus(property)]
+    fn set_interactive(&self, value: bool) -> zbus::Result<()>;
 }
