+++ rust/agama-lib/src/software/proxies/proposal.rs
@@ -18,10 +18,10 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! # D-Bus interface proxy for: `org.opensuse.Agama1.Locale`
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Software1.Proposal`
 //!
 //! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
-//! Source: `org.opensuse.Agama1.Locale.bus.xml`.
+//! Source: `org.opensuse.Agama.Software1.Proposal.bus.xml`.
 //!
 //! You may prefer to adapt it, instead of using it verbatim.
 //!
@@ -31,7 +31,6 @@
 //! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
 //! following zbus API can be used:
 //!
-//! * [`zbus::fdo::PeerProxy`]
 //! * [`zbus::fdo::PropertiesProxy`]
 //! * [`zbus::fdo::IntrospectableProxy`]
 //!
@@ -39,37 +38,41 @@
 //!
 //! [Writing a client proxy]: https://dbus2.github.io/zbus/client.html
 //! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
+
 use zbus::proxy;
 #[proxy(
-    default_service = "org.opensuse.Agama1",
-    interface = "org.opensuse.Agama1.Locale",
+    default_service = "org.opensuse.Agama.Software1",
+    interface = "org.opensuse.Agama.Software1.Proposal",
     assume_defaults = true
 )]
-pub trait Locale {
-    /// Commit method
-    fn commit(&self) -> zbus::Result<()>;
+pub trait Proposal {
+    /// AddResolvables method
+    fn add_resolvables(
+        &self,
+        Id: &str,
+        Type: u8,
+        Resolvables: &[&str],
+        Optional: bool,
+    ) -> zbus::Result<()>;
 
-    /// Keymap property
-    #[zbus(property)]
-    fn keymap(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_keymap(&self, value: &str) -> zbus::Result<()>;
+    /// GetResolvables method
+    fn get_resolvables(&self, Id: &str, Type: u8, Optional: bool) -> zbus::Result<Vec<String>>;
 
-    /// Locales property
-    #[zbus(property)]
-    fn locales(&self) -> zbus::Result<Vec<String>>;
-    #[zbus(property)]
-    fn set_locales(&self, value: &[&str]) -> zbus::Result<()>;
+    /// RemoveResolvables method
+    fn remove_resolvables(
+        &self,
+        Id: &str,
+        Type: u8,
+        Resolvables: &[&str],
+        Optional: bool,
+    ) -> zbus::Result<()>;
 
-    /// Timezone property
-    #[zbus(property)]
-    fn timezone(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_timezone(&self, value: &str) -> zbus::Result<()>;
-
-    /// UILocale property
-    #[zbus(property, name = "UILocale")]
-    fn uilocale(&self) -> zbus::Result<String>;
-    #[zbus(property, name = "UILocale")]
-    fn set_uilocale(&self, value: &str) -> zbus::Result<()>;
+    /// SetResolvables method
+    fn set_resolvables(
+        &self,
+        Id: &str,
+        Type: u8,
+        Resolvables: &[&str],
+        Optional: bool,
+    ) -> zbus::Result<()>;
 }
