+++ rust/agama-lib/src/software/proxies/software.rs
@@ -1,0 +1,105 @@
+// Copyright (c) [2024] SUSE LLC
+//
+// All Rights Reserved.
+//
+// This program is free software; you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by the Free
+// Software Foundation; either version 2 of the License, or (at your option)
+// any later version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+// more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, contact SUSE LLC.
+//
+// To contact SUSE LLC about this file by physical or electronic mail, you may
+// find current contact information at www.suse.com.
+
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Software1`
+//!
+//! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
+//! Source: `org.opensuse.Agama.Software1.bus.xml`.
+//!
+//! You may prefer to adapt it, instead of using it verbatim.
+//!
+//! More information can be found in the [Writing a client proxy] section of the zbus
+//! documentation.
+//!
+//! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
+//! following zbus API can be used:
+//!
+//! * [`zbus::fdo::PropertiesProxy`]
+//! * [`zbus::fdo::IntrospectableProxy`]
+//!
+//! Consequently `zbus-xmlgen` did not generate code for the above interfaces.
+//!
+//! [Writing a client proxy]: https://dbus2.github.io/zbus/client.html
+//! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
+
+use zbus::proxy;
+
+/// Software patterns map.
+///
+/// It uses the pattern name as key and a tuple containing the following information as value:
+///
+/// * Category.
+/// * Description.
+/// * Icon.
+/// * Summary.
+/// * Order.
+pub type PatternsMap = std::collections::HashMap<String, (String, String, String, String, String)>;
+
+#[proxy(
+    default_service = "org.opensuse.Agama.Software1",
+    default_path = "/org/opensuse/Agama/Software1",
+    interface = "org.opensuse.Agama.Software1",
+    assume_defaults = true
+)]
+pub trait Software1 {
+    /// AddPattern method
+    fn add_pattern(&self, id: &str) -> zbus::Result<bool>;
+
+    /// Finish method
+    fn finish(&self) -> zbus::Result<()>;
+
+    /// Install method
+    fn install(&self) -> zbus::Result<()>;
+
+    /// IsPackageAvailable method
+    fn is_package_available(&self, name: &str) -> zbus::Result<bool>;
+
+    /// IsPackageInstalled method
+    fn is_package_installed(&self, name: &str) -> zbus::Result<bool>;
+
+    /// ListPatterns method
+    fn list_patterns(&self, filtered: bool) -> zbus::Result<PatternsMap>;
+
+    /// Probe method
+    fn probe(&self) -> zbus::Result<()>;
+
+    /// Propose method
+    fn propose(&self) -> zbus::Result<()>;
+
+    /// ProvisionsSelected method
+    fn provisions_selected(&self, provisions: &[&str]) -> zbus::Result<Vec<bool>>;
+
+    /// RemovePattern method
+    fn remove_pattern(&self, id: &str) -> zbus::Result<bool>;
+
+    /// SetUserPatterns method
+    fn set_user_patterns(&self, add: &[&str], remove: &[&str]) -> zbus::Result<Vec<String>>;
+
+    /// UsedDiskSpace method
+    fn used_disk_space(&self) -> zbus::Result<String>;
+
+    /// ProbeFinished signal
+    #[zbus(signal)]
+    fn probe_finished(&self) -> zbus::Result<()>;
+
+    /// SelectedPatterns property
+    #[zbus(property)]
+    fn selected_patterns(&self) -> zbus::Result<std::collections::HashMap<String, u8>>;
+}
