+++ rust/agama-lib/src/storage/proxies/dasd/device.rs
@@ -18,10 +18,10 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! # D-Bus interface proxy for: `org.opensuse.Agama1.Locale`
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Storage1.DASD.Device`
 //!
 //! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
-//! Source: `org.opensuse.Agama1.Locale.bus.xml`.
+//! Source: `org.opensuse.Agama.Storage1.DASD.Device.bus.xml`.
 //!
 //! You may prefer to adapt it, instead of using it verbatim.
 //!
@@ -31,7 +31,6 @@
 //! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
 //! following zbus API can be used:
 //!
-//! * [`zbus::fdo::PeerProxy`]
 //! * [`zbus::fdo::PropertiesProxy`]
 //! * [`zbus::fdo::IntrospectableProxy`]
 //!
@@ -41,35 +40,44 @@
 //! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
 use zbus::proxy;
 #[proxy(
-    default_service = "org.opensuse.Agama1",
-    interface = "org.opensuse.Agama1.Locale",
+    default_service = "org.opensuse.Agama.Storage1",
+    interface = "org.opensuse.Agama.Storage1.DASD.Device",
     assume_defaults = true
 )]
-pub trait Locale {
-    /// Commit method
-    fn commit(&self) -> zbus::Result<()>;
+pub trait Device {
+    /// AccessType property
+    #[zbus(property)]
+    fn access_type(&self) -> zbus::Result<String>;
 
-    /// Keymap property
+    /// DeviceName property
     #[zbus(property)]
-    fn keymap(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_keymap(&self, value: &str) -> zbus::Result<()>;
+    fn device_name(&self) -> zbus::Result<String>;
 
-    /// Locales property
+    /// Diag property
     #[zbus(property)]
-    fn locales(&self) -> zbus::Result<Vec<String>>;
-    #[zbus(property)]
-    fn set_locales(&self, value: &[&str]) -> zbus::Result<()>;
+    fn diag(&self) -> zbus::Result<bool>;
 
-    /// Timezone property
+    /// Enabled property
     #[zbus(property)]
-    fn timezone(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_timezone(&self, value: &str) -> zbus::Result<()>;
+    fn enabled(&self) -> zbus::Result<bool>;
 
-    /// UILocale property
-    #[zbus(property, name = "UILocale")]
-    fn uilocale(&self) -> zbus::Result<String>;
-    #[zbus(property, name = "UILocale")]
-    fn set_uilocale(&self, value: &str) -> zbus::Result<()>;
+    /// Formatted property
+    #[zbus(property)]
+    fn formatted(&self) -> zbus::Result<bool>;
+
+    /// Id property
+    #[zbus(property)]
+    fn id(&self) -> zbus::Result<String>;
+
+    /// PartitionInfo property
+    #[zbus(property)]
+    fn partition_info(&self) -> zbus::Result<String>;
+
+    /// Status property
+    #[zbus(property)]
+    fn status(&self) -> zbus::Result<String>;
+
+    /// Type property
+    #[zbus(property)]
+    fn type_(&self) -> zbus::Result<String>;
 }
