+++ rust/agama-lib/src/storage/proxies/iscsi.rs
@@ -18,11 +18,8 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-mod software;
-pub use software::Software1Proxy;
+mod initiator;
+pub use initiator::InitiatorProxy;
 
-mod product;
-pub use product::{Product, ProductProxy as SoftwareProductProxy};
-
-mod proposal;
-pub use proposal::ProposalProxy;
+mod node;
+pub use node::NodeProxy;
