+++ rust/agama-lib/src/storage/proxies/iscsi/initiator.rs
@@ -18,10 +18,10 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! # D-Bus interface proxy for: `org.opensuse.Agama1.Locale`
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Storage1.ISCSI.Initiator`
 //!
 //! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
-//! Source: `org.opensuse.Agama1.Locale.bus.xml`.
+//! Source: `org.opensuse.Agama.Storage1.ISCSI.Initiator.bus.xml`.
 //!
 //! You may prefer to adapt it, instead of using it verbatim.
 //!
@@ -31,9 +31,9 @@
 //! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
 //! following zbus API can be used:
 //!
-//! * [`zbus::fdo::PeerProxy`]
-//! * [`zbus::fdo::PropertiesProxy`]
 //! * [`zbus::fdo::IntrospectableProxy`]
+//! * [`zbus::fdo::ObjectManagerProxy`]
+//! * [`zbus::fdo::PropertiesProxy`]
 //!
 //! Consequently `zbus-xmlgen` did not generate code for the above interfaces.
 //!
@@ -41,35 +41,29 @@
 //! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
 use zbus::proxy;
 #[proxy(
-    default_service = "org.opensuse.Agama1",
-    interface = "org.opensuse.Agama1.Locale",
+    default_service = "org.opensuse.Agama.Storage1",
+    interface = "org.opensuse.Agama.Storage1.ISCSI.Initiator",
     assume_defaults = true
 )]
-pub trait Locale {
-    /// Commit method
-    fn commit(&self) -> zbus::Result<()>;
+pub trait Initiator {
+    /// Delete method
+    fn delete(&self, node: &zbus::zvariant::ObjectPath<'_>) -> zbus::Result<u32>;
 
-    /// Keymap property
-    #[zbus(property)]
-    fn keymap(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_keymap(&self, value: &str) -> zbus::Result<()>;
+    /// Discover method
+    fn discover(
+        &self,
+        address: &str,
+        port: u32,
+        options: std::collections::HashMap<&str, &zbus::zvariant::Value<'_>>,
+    ) -> zbus::Result<u32>;
 
-    /// Locales property
-    #[zbus(property)]
-    fn locales(&self) -> zbus::Result<Vec<String>>;
-    #[zbus(property)]
-    fn set_locales(&self, value: &[&str]) -> zbus::Result<()>;
+    /// IBFT property
+    #[zbus(property, name = "IBFT")]
+    fn ibft(&self) -> zbus::Result<bool>;
 
-    /// Timezone property
+    /// InitiatorName property
     #[zbus(property)]
-    fn timezone(&self) -> zbus::Result<String>;
+    fn initiator_name(&self) -> zbus::Result<String>;
     #[zbus(property)]
-    fn set_timezone(&self, value: &str) -> zbus::Result<()>;
-
-    /// UILocale property
-    #[zbus(property, name = "UILocale")]
-    fn uilocale(&self) -> zbus::Result<String>;
-    #[zbus(property, name = "UILocale")]
-    fn set_uilocale(&self, value: &str) -> zbus::Result<()>;
+    fn set_initiator_name(&self, value: &str) -> zbus::Result<()>;
 }
