+++ rust/agama-lib/src/storage/proxies/iscsi/node.rs
@@ -18,10 +18,10 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! # D-Bus interface proxy for: `org.opensuse.Agama1.Locale`
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Storage1.ISCSI.Node`
 //!
 //! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
-//! Source: `org.opensuse.Agama1.Locale.bus.xml`.
+//! Source: `org.opensuse.Agama.Storage1.ISCSI.Node.bus.xml`.
 //!
 //! You may prefer to adapt it, instead of using it verbatim.
 //!
@@ -31,7 +31,6 @@
 //! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
 //! following zbus API can be used:
 //!
-//! * [`zbus::fdo::PeerProxy`]
 //! * [`zbus::fdo::PropertiesProxy`]
 //! * [`zbus::fdo::IntrospectableProxy`]
 //!
@@ -41,35 +40,47 @@
 //! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
 use zbus::proxy;
 #[proxy(
-    default_service = "org.opensuse.Agama1",
-    interface = "org.opensuse.Agama1.Locale",
+    default_service = "org.opensuse.Agama.Storage1",
+    interface = "org.opensuse.Agama.Storage1.ISCSI.Node",
     assume_defaults = true
 )]
-pub trait Locale {
-    /// Commit method
-    fn commit(&self) -> zbus::Result<()>;
+pub trait Node {
+    /// Login method
+    fn login(
+        &self,
+        options: std::collections::HashMap<&str, &zbus::zvariant::Value<'_>>,
+    ) -> zbus::Result<u32>;
 
-    /// Keymap property
-    #[zbus(property)]
-    fn keymap(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_keymap(&self, value: &str) -> zbus::Result<()>;
+    /// Logout method
+    fn logout(&self) -> zbus::Result<u32>;
 
-    /// Locales property
+    /// Address property
     #[zbus(property)]
-    fn locales(&self) -> zbus::Result<Vec<String>>;
-    #[zbus(property)]
-    fn set_locales(&self, value: &[&str]) -> zbus::Result<()>;
+    fn address(&self) -> zbus::Result<String>;
 
-    /// Timezone property
+    /// Connected property
     #[zbus(property)]
-    fn timezone(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_timezone(&self, value: &str) -> zbus::Result<()>;
+    fn connected(&self) -> zbus::Result<bool>;
 
-    /// UILocale property
-    #[zbus(property, name = "UILocale")]
-    fn uilocale(&self) -> zbus::Result<String>;
-    #[zbus(property, name = "UILocale")]
-    fn set_uilocale(&self, value: &str) -> zbus::Result<()>;
+    /// IBFT property
+    #[zbus(property, name = "IBFT")]
+    fn ibft(&self) -> zbus::Result<bool>;
+
+    /// Interface property
+    #[zbus(property)]
+    fn interface(&self) -> zbus::Result<String>;
+
+    /// Port property
+    #[zbus(property)]
+    fn port(&self) -> zbus::Result<u32>;
+
+    /// Startup property
+    #[zbus(property)]
+    fn startup(&self) -> zbus::Result<String>;
+    #[zbus(property)]
+    fn set_startup(&self, value: &str) -> zbus::Result<()>;
+
+    /// Target property
+    #[zbus(property)]
+    fn target(&self) -> zbus::Result<String>;
 }
