+++ rust/agama-lib/src/storage/proxies/zfcp/controller.rs
@@ -18,10 +18,10 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! # D-Bus interface proxy for: `org.opensuse.Agama1.Locale`
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Storage1.ZFCP.Controller`
 //!
 //! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
-//! Source: `org.opensuse.Agama1.Locale.bus.xml`.
+//! Source: `org.opensuse.Agama.Storage1.ZFCP.Controller.bus.xml`.
 //!
 //! You may prefer to adapt it, instead of using it verbatim.
 //!
@@ -31,7 +31,6 @@
 //! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
 //! following zbus API can be used:
 //!
-//! * [`zbus::fdo::PeerProxy`]
 //! * [`zbus::fdo::PropertiesProxy`]
 //! * [`zbus::fdo::IntrospectableProxy`]
 //!
@@ -41,35 +40,38 @@
 //! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
 use zbus::proxy;
 #[proxy(
-    default_service = "org.opensuse.Agama1",
-    interface = "org.opensuse.Agama1.Locale",
+    default_service = "org.opensuse.Agama.Storage1",
+    default_path = "/org/opensuse/Agama/Storage1",
+    interface = "org.opensuse.Agama.Storage1.ZFCP.Controller",
     assume_defaults = true
 )]
-pub trait Locale {
-    /// Commit method
-    fn commit(&self) -> zbus::Result<()>;
+pub trait Controller {
+    /// Activate method
+    fn activate(&self) -> zbus::Result<u32>;
 
-    /// Keymap property
-    #[zbus(property)]
-    fn keymap(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_keymap(&self, value: &str) -> zbus::Result<()>;
+    /// ActivateDisk method
+    fn activate_disk(&self, wwpn: &str, lun: &str) -> zbus::Result<u32>;
 
-    /// Locales property
-    #[zbus(property)]
-    fn locales(&self) -> zbus::Result<Vec<String>>;
-    #[zbus(property)]
-    fn set_locales(&self, value: &[&str]) -> zbus::Result<()>;
+    /// DeactivateDisk method
+    fn deactivate_disk(&self, wwpn: &str, lun: &str) -> zbus::Result<u32>;
 
-    /// Timezone property
-    #[zbus(property)]
-    fn timezone(&self) -> zbus::Result<String>;
-    #[zbus(property)]
-    fn set_timezone(&self, value: &str) -> zbus::Result<()>;
+    /// GetLUNs method
+    #[zbus(name = "GetLUNs")]
+    fn get_luns(&self, wwpn: &str) -> zbus::Result<Vec<String>>;
 
-    /// UILocale property
-    #[zbus(property, name = "UILocale")]
-    fn uilocale(&self) -> zbus::Result<String>;
-    #[zbus(property, name = "UILocale")]
-    fn set_uilocale(&self, value: &str) -> zbus::Result<()>;
+    /// GetWWPNs method
+    #[zbus(name = "GetWWPNs")]
+    fn get_wwpns(&self) -> zbus::Result<Vec<String>>;
+
+    /// Active property
+    #[zbus(property)]
+    fn active(&self) -> zbus::Result<bool>;
+
+    /// Channel property
+    #[zbus(property)]
+    fn channel(&self) -> zbus::Result<String>;
+
+    /// LUNScan property
+    #[zbus(property, name = "LUNScan")]
+    fn lunscan(&self) -> zbus::Result<bool>;
 }
