+++ rust/agama-lib/src/storage/proxies/zfcp/disk.rs
@@ -1,0 +1,44 @@
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Storage1.ZFCP.Disk`
+//!
+//! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
+//! Source: `org.opensuse.Agama.Storage1.ZFCP.Disk.bus.xml`.
+//!
+//! You may prefer to adapt it, instead of using it verbatim.
+//!
+//! More information can be found in the [Writing a client proxy] section of the zbus
+//! documentation.
+//!
+//! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
+//! following zbus API can be used:
+//!
+//! * [`zbus::fdo::PropertiesProxy`]
+//! * [`zbus::fdo::IntrospectableProxy`]
+//!
+//! Consequently `zbus-xmlgen` did not generate code for the above interfaces.
+//!
+//! [Writing a client proxy]: https://dbus2.github.io/zbus/client.html
+//! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
+use zbus::proxy;
+#[proxy(
+    default_service = "org.opensuse.Agama.Storage1",
+    default_path = "/org/opensuse/Agama/Storage1",
+    interface = "org.opensuse.Agama.Storage1.ZFCP.Disk",
+    assume_defaults = true
+)]
+pub trait Disk {
+    /// Channel property
+    #[zbus(property)]
+    fn channel(&self) -> zbus::Result<String>;
+
+    /// LUN property
+    #[zbus(property, name = "LUN")]
+    fn lun(&self) -> zbus::Result<String>;
+
+    /// Name property
+    #[zbus(property)]
+    fn name(&self) -> zbus::Result<String>;
+
+    /// WWPN property
+    #[zbus(property, name = "WWPN")]
+    fn wwpn(&self) -> zbus::Result<String>;
+}
