+++ rust/agama-lib/src/storage/proxies/zfcp/manager.rs
@@ -1,0 +1,36 @@
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Storage1.ZFCP.Manager`
+//!
+//! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
+//! Source: `org.opensuse.Agama.Storage1.DASD.Manager.bus.xml`.
+//!
+//! You may prefer to adapt it, instead of using it verbatim.
+//!
+//! More information can be found in the [Writing a client proxy] section of the zbus
+//! documentation.
+//!
+//! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
+//! following zbus API can be used:
+//!
+//! * [`zbus::fdo::PropertiesProxy`]
+//! * [`zbus::fdo::IntrospectableProxy`]
+//! * [`zbus::fdo::ObjectManagerProxy`]
+//!
+//! Consequently `zbus-xmlgen` did not generate code for the above interfaces.
+//!
+//! [Writing a client proxy]: https://dbus2.github.io/zbus/client.html
+//! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
+use zbus::proxy;
+#[proxy(
+    default_service = "org.opensuse.Agama.Storage1",
+    default_path = "/org/opensuse/Agama/Storage1",
+    interface = "org.opensuse.Agama.Storage1.ZFCP.Manager",
+    assume_defaults = true
+)]
+pub trait Manager {
+    /// Probe method
+    fn probe(&self) -> zbus::Result<()>;
+
+    /// AllowLUNScan property
+    #[zbus(property, name = "AllowLUNScan")]
+    fn allow_lunscan(&self) -> zbus::Result<bool>;
+}
