+++ rust/agama-lib/src/users/proxies.rs
@@ -18,10 +18,27 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! D-Bus interface proxies for: `org.opensuse.Agama.Users1.*`
+//! # D-Bus interface proxy for: `org.opensuse.Agama.Users1`
 //!
-//! This code was generated by `zbus-xmlgen` `3.1.0` from DBus introspection data.`.
-use zbus::dbus_proxy;
+//! This code was generated by `zbus-xmlgen` `5.0.0` from D-Bus introspection data.
+//! Source: `org.opensuse.Agama.Users1.bus.xml`.
+//!
+//! You may prefer to adapt it, instead of using it verbatim.
+//!
+//! More information can be found in the [Writing a client proxy] section of the zbus
+//! documentation.
+//!
+//! This type implements the [D-Bus standard interfaces], (`org.freedesktop.DBus.*`) for which the
+//! following zbus API can be used:
+//!
+//! * [`zbus::fdo::IntrospectableProxy`]
+//! * [`zbus::fdo::PropertiesProxy`]
+//!
+//! Consequently `zbus-xmlgen` did not generate code for the above interfaces.
+//!
+//! [Writing a client proxy]: https://dbus2.github.io/zbus/client.html
+//! [D-Bus standard interfaces]: https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces,
+use zbus::proxy;
 
 /// First user as it comes from D-Bus.
 ///
@@ -32,6 +49,7 @@ use zbus::dbus_proxy;
 /// * password
 /// * auto-login (enabled or not)
 /// * some optional and additional data
+// NOTE: Manually added to this file.
 pub type FirstUser = (
     String,
     String,
@@ -40,12 +58,13 @@ pub type FirstUser = (
     std::collections::HashMap<String, zbus::zvariant::OwnedValue>,
 );
 
-#[dbus_proxy(
-    interface = "org.opensuse.Agama.Users1",
+#[proxy(
     default_service = "org.opensuse.Agama.Manager1",
-    default_path = "/org/opensuse/Agama/Users1"
+    default_path = "/org/opensuse/Agama/Users1",
+    interface = "org.opensuse.Agama.Users1",
+    assume_defaults = true
 )]
-trait Users1 {
+pub trait Users1 {
     /// RemoveFirstUser method
     fn remove_first_user(&self) -> zbus::Result<u32>;
 
@@ -59,28 +78,28 @@ trait Users1 {
         user_name: &str,
         password: &str,
         auto_login: bool,
-        data: std::collections::HashMap<&str, zbus::zvariant::Value<'_>>,
+        data: std::collections::HashMap<&str, &zbus::zvariant::Value<'_>>,
     ) -> zbus::Result<(bool, Vec<String>)>;
 
     /// SetRootPassword method
     fn set_root_password(&self, value: &str, encrypted: bool) -> zbus::Result<u32>;
 
     /// SetRootSSHKey method
-    #[dbus_proxy(name = "SetRootSSHKey")]
+    #[zbus(name = "SetRootSSHKey")]
     fn set_root_sshkey(&self, value: &str) -> zbus::Result<u32>;
 
     /// Write method
     fn write(&self) -> zbus::Result<u32>;
 
     /// FirstUser property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn first_user(&self) -> zbus::Result<FirstUser>;
 
     /// RootPasswordSet property
-    #[dbus_proxy(property)]
+    #[zbus(property)]
     fn root_password_set(&self) -> zbus::Result<bool>;
 
     /// RootSSHKey property
-    #[dbus_proxy(property, name = "RootSSHKey")]
+    #[zbus(property, name = "RootSSHKey")]
     fn root_sshkey(&self) -> zbus::Result<String>;
 }
